+++ CHANGELOG.md
@@ -15,10 +15,12 @@ Synergy Wholesale WHMCS Domains Module
 ### Removed
 -
 
+
 ## 2.2.5 [Updated 11/08/2021]
 
 ### Fixed
 - Fixed Typo with Newline not getting interpreted
+- Fixed 'on sale' domain register pricing.
 
 ## 2.2.2 [Updated 12/03/2021]